lib refactoring

+++ src/main/java/com/vendo/auth_service/application/otp/service/EmailOtpService.java
@@ -1,12 +1,14 @@
 package com.vendo.auth_service.application.otp.service;
 
+import com.vendo.auth_service.application.otp.common.exception.InvalidOtpException;
+import com.vendo.auth_service.application.otp.common.exception.OtpAlreadySentException;
 import com.vendo.auth_service.domain.otp.OtpPolicyService;
 import com.vendo.auth_service.port.otp.OtpEmailNotificationPort;
 import com.vendo.auth_service.port.otp.OtpGenerator;
 import com.vendo.auth_service.port.otp.OtpStorage;
 import com.vendo.auth_service.adapter.otp.out.props.OtpNamespace;
-import com.vendo.integration.kafka.event.EmailOtpEvent;
-import com.vendo.integration.redis.common.exception.OtpExpiredException;
+import com.vendo.event_lib.EmailOtpEvent;
+import com.vendo.redis_lib.exception.OtpExpiredException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -16,7 +18,6 @@ import java.util.Optional;
 @Slf4j
 @Service
 @RequiredArgsConstructor
-// TODO write mocking tests
 public class EmailOtpService {
 
     private final OtpStorage otpStorage;
@@ -28,14 +29,17 @@ public class EmailOtpService {
     private final OtpEmailNotificationPort otpEmailNotificationPort;
 
     public void sendOtp(EmailOtpEvent event, OtpNamespace otpNamespace) {
-        otpPolicyService.checkIfInactive(otpStorage.hasActiveKey(otpNamespace.getEmail().buildPrefix(event.getEmail())));
+        boolean activeKey = otpStorage.hasActiveKey(otpNamespace.getEmail().buildPrefix(event.getEmail()));
+        if (activeKey) {
+            throw new OtpAlreadySentException("Otp has already sent.");
+        }
 
         String otp = otpGenerator.generate();
-        event.setOtp(otp);
 
         otpStorage.saveValue(otpNamespace.getOtp().buildPrefix(otp), event.getEmail(), otpNamespace.getOtp().getTtl());
         otpStorage.saveValue(otpNamespace.getEmail().buildPrefix(event.getEmail()), otp, otpNamespace.getEmail().getTtl());
 
+        event.setOtp(otp);
         otpEmailNotificationPort.sendOtpEmailNotification(event);
     }
 
@@ -51,24 +55,34 @@ public class EmailOtpService {
         otpEmailNotificationPort.sendOtpEmailNotification(event);
     }
 
-    public String verifyOtpAndConsume(String otp, String expectedEmail, OtpNamespace namespace) {
+    public void verifyEmailByOtp(String otp, String expectedEmail, OtpNamespace namespace) {
         String actualEmail = otpStorage.getValue(namespace.getOtp().buildPrefix(otp))
                 .orElseThrow(() -> new OtpExpiredException("Otp session expired."));
 
-        otpPolicyService.checkIfEmailMatches(expectedEmail, actualEmail);
+        if (expectedEmail != null && !expectedEmail.equals(actualEmail)) {
+            throw new InvalidOtpException("Invalid otp.");
+        }
 
         otpStorage.deleteValues(
                 namespace.getOtp().buildPrefix(otp),
                 namespace.getEmail().buildPrefix(actualEmail),
                 namespace.getAttempts().buildPrefix(actualEmail)
         );
+    }
 
-        return actualEmail;
+    public String verifyByOtp(String otp, OtpNamespace namespace) {
+        String email = otpStorage.getValue(namespace.getOtp().buildPrefix(otp))
+                .orElseThrow(() -> new OtpExpiredException("Otp session expired."));
+
+        otpStorage.deleteValues(namespace.getOtp().buildPrefix(otp));
+
+        return email;
     }
 
     private void increaseResendAttemptsOrThrow(String email, OtpNamespace otpNamespace) {
         Optional<String> attempts = otpStorage.getValue(otpNamespace.getAttempts().buildPrefix(email));
-        int attempt = otpPolicyService.checkAndIncreaseAttempts(attempts.map(Integer::parseInt).orElse(0));
+        int attempt = otpPolicyService.throwOrIncreaseAttempts(attempts.map(Integer::parseInt).orElse(0));
+
         otpStorage.saveValue(
                 otpNamespace.getAttempts().buildPrefix(email),
                 String.valueOf(attempt),
@@ -79,8 +93,9 @@ public class EmailOtpService {
         Optional<String> otp = otpStorage.getValue(otpNamespace.getEmail().buildPrefix(email));
 
         if (otp.isEmpty()) {
-            otp = otp.map(o -> otpGenerator.generate());
-            otpStorage.saveValue(otpNamespace.getEmail().buildPrefix(email), otp.get(), otpNamespace.getOtp().getTtl());
+            String newOtp = otpGenerator.generate();
+            otpStorage.saveValue(otpNamespace.getEmail().buildPrefix(email), newOtp, otpNamespace.getOtp().getTtl());
+            return newOtp;
         }
 
         return otp.get();
